+++ src/commonMain/kotlin/uk/gov/hmrc/calculator/model/bands/TaxFreeAllowance.kt
@@ -16,9 +16,11 @@
 package uk.gov.hmrc.calculator.model.bands
 
 import uk.gov.hmrc.calculator.model.TaxYear
+import kotlin.jvm.JvmSynthetic
 
 internal object TaxFreeAllowance {
 
+    @JvmSynthetic
     fun getAllowance(taxYear: TaxYear): Double =
         when (taxYear) {
             TaxYear.TWENTY_TWENTY -> 12500.00