+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Removed `standardLifetimeAllowance` for Pension validation.
 - Allow Pension to go above `annualAllowance`.
 - Fix Student Loan internal calculation to use yearly wage instead of deducting pension.
+- Fix internal object so they're not accessible by jvm consumer.
 
 ## [2.10.2] - 2024-03-08Z